backend/requirements.txt
channels==4.2.0
Django==4.2.16
django_celery_beat==2.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
python-dotenv==1.0.1
uvicorn
uvicorn[standard]
watchdog
django-cors-headers
django-allauth
django_celery_results
psycopg2-binary
requests
pillow
redis>=3.5.0,<5.0
channels_redis	